requirements.txt
asgiref==3.3.1
Django==3.1.5
gunicorn==20.0.4
python-dotenv==0.15.0
pytz==2020.5
sqlparse==0.4.1
APScheduler==3.7.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
configparser==5.0.1
cryptography==3.3.1
dj-database-url==0.5.0
django-apscheduler==0.5.1
django-heroku==0.3.1
django-livereload-server==0.3.2
django-rest-knox==4.1.0
djangorestframework==3.12.2
idna==2.10
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
requests==2.25.1
six==1.15.0
soupsieve==2.1
tornado==6.1
twilio==6.51.0
tzlocal==2.1
urllib3==1.26.2
whitenoise==5.2.0